second/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
pika
bcrypt
python-dotenv